+++ R/Class_tskrrHeterogenousImpute.R
@@ -4,7 +4,7 @@
 #' class \code{\link[xnet:tskrrHeterogenous-class]{tskrrHeterogenous}} and
 #' \code{\link[xnet:tskrrImpute-class]{tskrrImpute}}
 #' specifically for heterogenous networks with imputed values. It is
-#' the result of the function \code{\link{impute_tskrr}} or the function \code{\link{impute_loo}}.
+#' the result of the function \code{\link{impute_tskrr}}.
 #'
 #' @slot y the matrix with responses
 #' @slot k the eigen decomposition of the kernel matrix for the rows
@@ -22,12 +22,6 @@
 #' are integers indicating the row resp column number.
 #' @slot imputeid a vector with integer values indicating which of
 #' the values in \code{y} are imputed
-#' @slot imputemethod a character value with either \code{"predictions"} or
-#' "loo" indicating whether the predictions or the leave-one-out
-#' crossvalidation values were used in the imputation.
-#' @slot loo_settings in case \code{imputemethod = "loo"}, a list with
-#' two elements named \code{exclusion} and \code{replaceby0}. These
-#' indicate the arguments used in \code{\link{get_loo_fun}}.
 #' @slot niter an integer value gving the number of iterations used
 #' @slot tol a numeric value with the tolerance used
 #'